requirements.txt
wheel
requests
python-dotenv
pytest
pyflac
psycopg2-binary